requirements.txt
fastapi
uvicorn[standard]
python-multipart
sqlalchemy
psycopg2-binary
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
llama-index-core
llama-index-llms-ollama
llama-index-embeddings-ollama
llama-index-embeddings-huggingface
llama-index-vector-stores-postgres
llama-index-readers-file
pymupdf
openpyxl
pandas
sentence-transformers
unstructured
pydantic
bcrypt


